+++ src/ResourceManager/Storage/Commands.Management.Storage.Test/ScenarioTests/StorageAccountTests.cs
@@ -121,5 +121,26 @@ namespace Microsoft.Azure.Commands.Management.Storage.Test.ScenarioTests
         {
             TestController.NewInstance.RunPsTest(_logger, "Test-SetAzureStorageAccountStorageV2");
         }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestGetAzureStorageSubscriptionUsage()
+        {
+            TestController.NewInstance.RunPsTest(_logger, "Test-GetAzureStorageSubscriptionUsage");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestGetAzureStorageLocationUsage()
+        {
+            TestController.NewInstance.RunPsTest(_logger, "Test-GetAzureStorageLocationUsage");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestAzureStorageAccountManagementPolicy()
+        {
+            TestController.NewInstance.RunPsTest(_logger, "Test-AzureStorageAccountManagementPolicy");
+        }
     }
 }