+++ src/ResourceManager/Storage/Commands.Management.Storage/ChangeLog.md
@@ -27,6 +27,35 @@
     - New-AzureRmStorageAccount
     - Set-AzureRmStorageAccount
 
+## Version 5.0.4-preview
+* Support create Storage account with Hierarchical Namespace enabled 
+       - New-AzureRmStorageAccount
+
+## Version 5.0.3-preview
+* Support Management Policy in Storage
+       - Remove-AzureRmStorageAccountNetworkRule
+	   - Set-AzureRmStorageAccountManagementPolicy
+	   - Get-AzureRmStorageAccountManagementPolicy
+	   - Remove-AzureRmStorageAccountManagementPolicy
+
+## Version 5.0.2-preview
+* Support Immutability Policy in AzureRm.Storage
+        - Remove-AzureRmStorageAccountNetworkRule,
+        - Get-AzureRmStorageContainer,
+        - Update-AzureRmStorageContainer,
+        - New-AzureRmStorageContainer,
+        - Remove-AzureRmStorageContainer,
+        - Add-AzureRmStorageContainerLegalHold,
+        - Remove-AzureRmStorageContainerLegalHold,
+        - Set-AzureRmStorageContainerImmutabilityPolicy,
+        - Get-AzureRmStorageContainerImmutabilityPolicy,
+        - Remove-AzureRmStorageContainerImmutabilityPolicy,
+        - Lock-AzureRmStorageContainerImmutabilityPolicy
+
+## Version 5.0.1-preview
+* Support get the Storage resource usage of a specific location
+    - Get-AzureRmStorageUsage
+
 ## Version 5.0.0
 * Set minimum dependency of module to PowerShell 5.0
 * Introduced multiple breaking changes